Create dedicated dates module
