+++ packages/tendermint-rpc/src/adaptors/v0-33/hasher.spec.ts
@@ -1,7 +1,7 @@
 import { fromBase64, fromHex } from "@cosmjs/encoding";
 import { ReadonlyDate } from "readonly-date";
 
-import { ReadonlyDateWithNanoseconds } from "../../types";
+import { ReadonlyDateWithNanoseconds } from "../../dates";
 import { hashBlock, hashTx } from "./hasher";
 
 describe("Hasher", () => {
