+++ packages/tendermint-rpc/src/index.ts
@@ -1,7 +1,7 @@
 export { Adaptor } from "./adaptor";
 export { adaptor33, adaptor34 } from "./adaptors";
 export { Client } from "./client";
-export { DateTime } from "./encodings";
+export { DateTime, ReadonlyDateWithNanoseconds } from "./dates";
 export {
   AbciInfoRequest,
   AbciQueryParams,
@@ -72,10 +72,4 @@ export {
   VoteType,
 } from "./responses";
 export { HttpClient, WebsocketClient } from "./rpcclients"; // TODO: Why do we export those outside of this package?
-export {
-  BlockIdFlag,
-  CommitSignature,
-  ReadonlyDateWithNanoseconds,
-  ValidatorEd25519Pubkey,
-  ValidatorPubkey,
-} from "./types";
+export { BlockIdFlag, CommitSignature, ValidatorEd25519Pubkey, ValidatorPubkey } from "./types";
